move instrumentation code into separate file
