+++ core/instrumentation.py
@@ -1,0 +1,24 @@
+from opentelemetry import trace
+from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
+from opentelemetry.sdk.resources import Resource
+import logging
+
+
+class SpanFormatter(logging.Formatter):
+    def format(self, record):
+        trace_id = trace.get_current_span().get_span_context().trace_id
+        if trace_id == 0:
+            record.trace_id = None
+        else:
+            record.trace_id = "{trace:032x}".format(trace=trace_id)
+        return super().format(record)
+
+
+def build_instrumentation(hostname: str, port: int):
+    resource = Resource(attributes={"service.name": "ss-backend"})
+
+    trace.set_tracer_provider(TracerProvider(resource=resource))
+    otlp_exporter = OTLPSpanExporter(endpoint=f"{hostname}:{port}", insecure=True)
+    trace.get_tracer_provider().add_span_processor(BatchSpanProcessor(otlp_exporter))
